move focus trap to its own file
