+++ packages/components/src/Modal/Modal.tsx
@@ -1,10 +1,11 @@
-import React, { ReactNode, RefObject, useEffect, useRef } from "react";
+import React, { ReactNode, RefObject, useEffect } from "react";
 import ReactDOM from "react-dom";
 import classnames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
 import { useOnKeyDown } from "@jobber/hooks";
 import styles from "./Modal.css";
 import sizes from "./Sizes.css";
+import { useFocusTrap } from "./useFocusTrap";
 import { Typography } from "../Typography";
 import { Button, ButtonProps } from "../Button";
 import { ButtonDismiss } from "../ButtonDismiss";
@@ -96,51 +97,6 @@ export function Modal({
       onRequestClose();
     }
   }
-}
-
-function useFocusTrap<T extends HTMLElement>() {
-  const ref = useRef<T>();
-  const focusables = [
-    "button",
-    "[href]",
-    "input",
-    "select",
-    "textarea",
-    '[tabindex]:not([tabindex="-1"])',
-  ];
-
-  function handleKeyDown(event: KeyboardEvent) {
-    if (!ref.current || event.key !== "Tab") {
-      return;
-    }
-
-    const elements = ref.current.querySelectorAll<HTMLElement>(
-      focusables.join(", "),
-    );
-    const firstElement = ref.current;
-    const lastElement = elements[elements.length - 1];
-
-    if (event.shiftKey) {
-      if (document.activeElement === firstElement) {
-        lastElement.focus();
-        event.preventDefault();
-      }
-    } else {
-      if (document.activeElement === lastElement) {
-        firstElement.focus();
-        event.preventDefault();
-      }
-    }
-  }
-
-  useEffect(() => {
-    ref.current?.addEventListener("keydown", handleKeyDown);
-    return () => {
-      ref.current?.removeEventListener("keydown", handleKeyDown);
-    };
-  });
-
-  return ref;
 }
 
 interface HeaderProps {
